Loggerhead Sea Turtle (Caretta caretta) Occurrences in the Mediterranean Sea

This dataset contains a subset of global occurrences of the Loggerhead Sea Turtle (Caretta caretta), filtered for records in the Mediterranean Sea. The data were sourced from the Global Biodiversity Information Facility (GBIF).

Usage

data("caretta")

Format

A data frame with 6785 rows and 5 columns:

decimalLongitude

Numeric. Longitude coordinates (WGS84).

decimalLatitude

Numeric. Latitude coordinates (WGS84).

year

Integer. The year in which the occurrence was recorded.

species

Character. The scientific name of the species, i.e., Caretta caretta.

coordinateUncertaintyInMeters

Numeric. The uncertainty of the coordinates in meters.

Source

Global Biodiversity Information Facility (GBIF), https://www.gbif.org/species/8894817

Details

The dataset has been filtered to include only records within the Mediterranean Sea. The occurrence data cover multiple years, which provides information on the temporal distribution of the species in this region.